+++ sqlglot/optimizer/qualify_tables.py
