+++ tests/test_optimizer.py
@@ -2,7 +2,8 @@ import doctest
 import os
 import unittest
 
-from sqlglot import optimizer
+from sqlglot.optimizer.qualify_tables import qualify_tables
+from sqlglot.optimizer.qualify_columns import qualify_columns
 from sqlglot import parse_one
 from sqlglot.errors import OptimizeError
 from tests.helpers import load_sql_fixture_pairs, load_sql_fixtures
@@ -12,9 +13,9 @@ def load_tests(loader, tests, ignore):  # pylint: disable=unused-argument
     """
     This finds and runs all the doctests in the expressions module
     """
-    from sqlglot.optimizer.rules import qualify_columns
+    from sqlglot.optimizer import qualify_columns as mod
 
-    tests.addTests(doctest.DocTestSuite(qualify_columns))
+    tests.addTests(doctest.DocTestSuite(mod))
     return tests
 
 
@@ -24,23 +25,19 @@ class TestOptimizer(unittest.TestCase):
 
     def test_qualify_tables(self):
         self.assertEqual(
-            optimizer.qualify_tables(parse_one("SELECT 1 FROM z"), db="db").sql(),
+            qualify_tables(parse_one("SELECT 1 FROM z"), db="db").sql(),
             "SELECT 1 FROM db.z",
         )
         self.assertEqual(
-            optimizer.qualify_tables(
-                parse_one("SELECT 1 FROM z"), db="db", catalog="c"
-            ).sql(),
+            qualify_tables(parse_one("SELECT 1 FROM z"), db="db", catalog="c").sql(),
             "SELECT 1 FROM c.db.z",
         )
         self.assertEqual(
-            optimizer.qualify_tables(
-                parse_one("SELECT 1 FROM y.z"), db="db", catalog="c"
-            ).sql(),
+            qualify_tables(parse_one("SELECT 1 FROM y.z"), db="db", catalog="c").sql(),
             "SELECT 1 FROM c.y.z",
         )
         self.assertEqual(
-            optimizer.qualify_tables(
+            qualify_tables(
                 parse_one("SELECT 1 FROM x.y.z"), db="db", catalog="c"
             ).sql(),
             "SELECT 1 FROM x.y.z",
@@ -54,7 +51,7 @@ class TestOptimizer(unittest.TestCase):
         for sql, expected in load_sql_fixture_pairs("optimizer/qualify_columns.sql"):
             with self.subTest(sql):
                 self.assertEqual(
-                    optimizer.qualify_columns(parse_one(sql), schema=schema).sql(),
+                    qualify_columns(parse_one(sql), schema=schema).sql(),
                     expected,
                 )
 
@@ -66,4 +63,4 @@ class TestOptimizer(unittest.TestCase):
         for sql in load_sql_fixtures("optimizer/qualify_columns__invalid.sql"):
             with self.subTest(sql):
                 with self.assertRaises(OptimizeError):
-                    optimizer.qualify_columns(parse_one(sql), schema=schema)
+                    qualify_columns(parse_one(sql), schema=schema)
